Outsource new db layer into a library project
