+++ androidrepository/src/main/java/net/syncthing/repository/android/database/converters/DateConverter.kt
@@ -1,0 +1,12 @@
+package net.syncthing.repository.android.database.converters
+
+import android.arch.persistence.room.TypeConverter
+import java.util.*
+
+class DateConverter {
+    @TypeConverter
+    fun toLong(date: Date) = date.time
+
+    @TypeConverter
+    fun fromLong(time: Long) = Date(time)
+}
