+++ app/build.gradle
@@ -1,6 +1,5 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-kapt'
 apply plugin: 'com.github.ben-manes.versions'
 
 android {
@@ -49,16 +48,9 @@ android {
     dataBinding {
         enabled = true
     }
-    kapt {
-        arguments {
-            arg("room.schemaLocation", "$projectDir/schemas".toString())
-        }
-    }
 }
 
 dependencies {
-    def room_version = "1.1.1"
-
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation "org.jetbrains.anko:anko-commons:$anko_version"
     implementation "org.jetbrains.anko:anko-coroutines:$anko_version"
@@ -88,6 +80,5 @@ dependencies {
     implementation 'com.google.zxing:core:3.3.0'
     implementation 'com.github.apl-devs:appintro:v4.2.3'
 
-    implementation "android.arch.persistence.room:runtime:$room_version"
-    kapt "android.arch.persistence.room:compiler:$room_version"
+    implementation project(':androidrepository')
 }
