--- app/src/main/kotlin/net/syncthing/lite/library/repository/database/converters/DateConverter.kt
@@ -1,12 +1,0 @@
-package net.syncthing.lite.library.repository.database.converters
-
-import android.arch.persistence.room.TypeConverter
-import java.util.*
-
-class DateConverter {
-    @TypeConverter
-    fun toLong(date: Date) = date.time
-
-    @TypeConverter
-    fun fromLong(time: Long) = Date(time)
-}
