--- app/src/main/kotlin/net/syncthing/lite/library/repository/database/converters/DeviceIdConverter.kt
@@ -1,12 +1,0 @@
-package net.syncthing.lite.library.repository.database.converters
-
-import android.arch.persistence.room.TypeConverter
-import net.syncthing.java.core.beans.DeviceId
-
-class DeviceIdConverter {
-    @TypeConverter
-    fun fromString(deviceId: String) = DeviceId(deviceId)
-
-    @TypeConverter
-    fun toString(deviceId: DeviceId) = deviceId.deviceId
-}
